fix selection macro

+++ Tests/StructuredQueriesTests/DeleteTests.swift
@@ -9,13 +9,14 @@ extension SnapshotTests {
     @Test func deleteAll() throws {
       try assertQuery(Reminder.delete().returning(\.self)) {
         """
-        DELETE FROM "reminders" RETURNING "reminders"."id", "reminders"."date", "reminders"."isCompleted", "reminders"."isFlagged", "reminders"."notes", "reminders"."priority", "reminders"."remindersListID", "reminders"."title"
+        DELETE FROM "reminders" RETURNING "reminders"."id", "reminders"."assignedUserID", "reminders"."date", "reminders"."isCompleted", "reminders"."isFlagged", "reminders"."notes", "reminders"."priority", "reminders"."remindersListID", "reminders"."title"
         """
       }results: {
         #"""
         ┌─────────────────────────────────────────┐
         │ Reminder(                               │
         │   id: 1,                                │
+        │   assignedUserID: 1,                    │
         │   date: Date(2001-01-01T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: false,                     │
@@ -27,6 +28,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 2,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2000-12-30T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: true,                      │
@@ -38,6 +40,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 3,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2001-01-01T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: false,                     │
@@ -49,6 +52,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 4,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2000-06-25T00:00:00.000Z), │
         │   isCompleted: true,                    │
         │   isFlagged: false,                     │
@@ -60,6 +64,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 5,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2001-01-01T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: false,                     │
@@ -71,6 +76,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 6,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2001-01-03T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: true,                      │
@@ -82,6 +88,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 7,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2000-12-30T00:00:00.000Z), │
         │   isCompleted: true,                    │
         │   isFlagged: false,                     │
@@ -93,6 +100,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 8,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2001-01-05T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: false,                     │
@@ -104,6 +112,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 9,                                │
+        │   assignedUserID: nil,                  │
         │   date: Date(2001-01-03T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: false,                     │
@@ -119,6 +128,7 @@ extension SnapshotTests {
         ├─────────────────────────────────────────┤
         │ Reminder(                               │
         │   id: 10,                               │
+        │   assignedUserID: nil,                  │
         │   date: Date(2000-12-30T00:00:00.000Z), │
         │   isCompleted: true,                    │
         │   isFlagged: false,                     │
@@ -146,13 +156,14 @@ extension SnapshotTests {
     @Test func deleteID1() throws {
       try assertQuery(Reminder.delete().where { $0.id == 1 }.returning(\.self)) {
         """
-        DELETE FROM "reminders" WHERE ("reminders"."id" = 1) RETURNING "reminders"."id", "reminders"."date", "reminders"."isCompleted", "reminders"."isFlagged", "reminders"."notes", "reminders"."priority", "reminders"."remindersListID", "reminders"."title"
+        DELETE FROM "reminders" WHERE ("reminders"."id" = 1) RETURNING "reminders"."id", "reminders"."assignedUserID", "reminders"."date", "reminders"."isCompleted", "reminders"."isFlagged", "reminders"."notes", "reminders"."priority", "reminders"."remindersListID", "reminders"."title"
         """
-      } results: {
+      }results: {
         """
         ┌─────────────────────────────────────────┐
         │ Reminder(                               │
         │   id: 1,                                │
+        │   assignedUserID: 1,                    │
         │   date: Date(2001-01-01T00:00:00.000Z), │
         │   isCompleted: false,                   │
         │   isFlagged: false,                     │
